QItemViews: hide ToolTip/What's this when cell has no data for it
A QToolTip/QWhatsThis was not hidden when the cursor moved to a cell
which does not return valid data for Qt::ToolTip/WhatsThisRole or when
the index is not valid. Therefore a wrong information was shown e.g.
when the cursor moved from a cell with a tooltip to one without.
Fix it by passing an empty string to QToolTip/QWhatsThis::showText().
This syncs the behavior with QGraphicsScene::helpEvent().

Fixes: QTBUG-78722
Change-Id: Ie99fe3b1d35d2f5be41dd65e2fe3173b0cc551b2
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>